requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
alembic>=1.13.3
openpyxl>=3.1.5
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic[email]>=2.10.0
# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
